requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML
anthropic==0.25.0
langgraph==0.0.50
langchain==0.1.20
langchain-core==0.1.52

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# API
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Utilities
httpx==0.25.2
tenacity==8.2.3
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.1
mypy==1.7.1
ruff==0.1.9
